requirements.txt
pandas==0.22.0
psycopg2-binary==2.7.4
pytest==3.6.1
